Made num_preceding_reads an int64_t.

+++ src/concurrency/fifo_enforcer.cc
@@ -84,7 +84,7 @@ void fifo_enforcer_sink_t::pump() THROWS_NOTHING {
         writer_queue_t::iterator it =
             waiting_writers.find(transition_timestamp_t::starting_from(state.timestamp));
         if (it != waiting_writers.end()) {
-            int expected_num_reads = it->second.first;
+            int64_t expected_num_reads = it->second.first;
             rassert(state.num_reads <= expected_num_reads);
 
             if (state.num_reads == expected_num_reads) {
@@ -109,7 +109,7 @@ void fifo_enforcer_sink_t::finish_a_reader(DEBUG_ONLY_VAR state_timestamp_t time
     state.num_reads++;
 }
 
-void fifo_enforcer_sink_t::finish_a_writer(transition_timestamp_t timestamp, DEBUG_ONLY_VAR int num_preceding_reads) THROWS_NOTHING {
+void fifo_enforcer_sink_t::finish_a_writer(transition_timestamp_t timestamp, DEBUG_ONLY_VAR int64_t num_preceding_reads) THROWS_NOTHING {
     assert_thread();
 
     rassert(state.timestamp == timestamp.timestamp_before());
